requirements.txt
Django==4.1.7
django-environ
psycopg2
djangorestframework
djangorestframework-jsonapi
django-filter
Pillow